code aufgeräumt und maingamecontrollers setupSpielfeld zu setupsspiel umbenannt da irreführend sein kann
