+++ src/main/java/com/turksat46/schiffgehtunter/CreateGameController.java
@@ -66,7 +66,7 @@ public class CreateGameController {
 
         mainGameController = fxmlLoader.getController();
         mainGameController.setCurrentMode(cb2.getSelectionModel().getSelectedIndex());
-        mainGameController.setupSpielfeld((int)groesseslider.getValue(), stage);
+        mainGameController.setupSpiel((int)groesseslider.getValue(), stage);
         Stage thisstage = (Stage) cb.getScene().getWindow();
         thisstage.close();
     }
